# springboot-jwt-api/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: School Management System API
  description: |-
    REST API for the School Management System — authentication, user management, master data, enrollments, grades, push notifications (Firebase Cloud Messaging) and audit log.

    Sign in with POST /api/v1/auth/login (admin/admin@123), then click Authorize and paste the accessToken.
  contact:
    name: Sothea
    email: keoleaphy@gmail.com
  version: v1.0
servers:
- url: http://localhost:30033
  description: Generated server url
security:
- bearerAuth: []
...
paths:
...
  /api/v1/notifications/{id}/read:
    put:
      tags:
      - Notifications
      summary: Mark one notification as read
      description: '**Requires permission:** `notification.read`'
      operationId: markRead
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseNotificationResponse"
...
  /api/v1/notifications/send:
    post:
      tags:
      - Notifications
      summary: Send a push notification
      description: Broadcast to every active user of a role (e.g. ROLE_STUDENT) or
        to an explicit list of userIds **Requires permission:** `notification.send`
      operationId: send
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendNotificationRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSendResult"
...
  /api/v1/devices:
    get:
      tags:
      - Devices
      summary: List my registered devices
      operationId: myDevices
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListDeviceResponse"
    post:
      tags:
      - Devices
      summary: Register this device for push
      description: Stores the FCM registration token of the signed-in user's device.
        Call after login and whenever Firebase rotates the token.
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDeviceRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseDeviceResponse"
    delete:
      tags:
      - Devices
      summary: Unregister a device
      description: Stops push delivery to the given token
      operationId: unregister
      parameters:
      - name: token
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
...
  /api/v1/notifications:
    get:
      tags:
      - Notifications
      summary: My notifications
      description: Paginated inbox; pass unreadOnly=true to filter **Requires permission:**
        `notification.read`
      operationId: inbox
      parameters:
      - name: unreadOnly
        in: query
        required: false
        schema:
          type: boolean
          default: false
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: sort
        in: query
        required: false
        schema:
          type: string
          default: "id,desc"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseNotificationResponse"
...
components:
  schemas:
...
    ApiResponseNotificationResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          $ref: "#/components/schemas/NotificationResponse"
    NotificationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        body:
          type: string
        type:
          type: string
        entityType:
          type: string
        entityId:
          type: string
        read:
          type: boolean
        sentStatus:
          type: string
        createdAt:
          type: string
          format: date-time
...
    SendNotificationRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 150
          minLength: 0
        body:
          type: string
          maxLength: 500
          minLength: 0
        role:
          type: string
        userIds:
          type: array
          items:
            type: integer
            format: int32
      required:
      - body
      - title
    ApiResponseSendResult:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          $ref: "#/components/schemas/SendResult"
    SendResult:
      type: object
      properties:
        recipients:
          type: integer
          format: int32
        devicesReached:
          type: integer
          format: int32
        pushEnabled:
          type: boolean
...
    RegisterDeviceRequest:
      type: object
      properties:
        token:
          type: string
          minLength: 1
        deviceType:
          type: string
        deviceName:
          type: string
      required:
      - token
    ApiResponseDeviceResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          $ref: "#/components/schemas/DeviceResponse"
    DeviceResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        token:
          type: string
        deviceType:
          type: string
        deviceName:
          type: string
        status:
          type: string
...
    ApiResponseVoid:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data: {}
...
    ApiResponsePageResponseNotificationResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          $ref: "#/components/schemas/PageResponseNotificationResponse"
    PageResponseNotificationResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/NotificationResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    ApiResponseListDeviceResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/DeviceResponse"
...
  securitySchemes:
    bearerAuth:
      type: http
      description: Paste the access token from POST /api/v1/auth/login
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT